Http proxy

+++ src/common/http/httpc.cpp
@@ -251,7 +251,7 @@ optional<Error> HttpClient::send_request(const char *host, Connection *conn, Req
                     if (written == 0) {
                         has_more = false;
                     } else {
-                        log_debug(httpc, "out: %*s", written, reinterpret_cast<char *>(buffer));
+                        log_debug(httpc, "out: %*s", static_cast<int>(written), reinterpret_cast<char *>(buffer));
                     }
                     return written;
                 } else {
